requirements.txt
absl-py==2.2.2
astroid==3.3.9
bandit==1.8.3
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
distlib==0.3.9
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
grpcio==1.71.0
identify==2.6.9
idna==3.10
isort==6.0.1
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy==1.15.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pbr==6.1.1
pillow==11.1.0
platformdirs==4.3.7
pre_commit==4.2.0
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
Pygments==2.19.1
pylint==3.3.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.25.1
six==1.17.0
stevedore==5.4.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
types-toml==0.10.8.20240310
typing_extensions==4.13.1
tzdata==2025.2
ultralytics==8.3.107
ultralytics-thop==2.0.14
urllib3==2.4.0
virtualenv==20.30.0
Werkzeug==3.1.3